% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharrell.R
\name{pharrell}
\alias{pharrell}
\title{Pharrell Data Envelopment Analysis Gadget}
\usage{
pharrell()
}
\value{
A data frame of the efficiency results that include efficiency scores and peers' determination.
}
\description{
Pharrell Data Envelopment Analysis Gadget
}
