% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrell.R
\name{farrell}
\alias{farrell}
\title{farrell Data Envelopment Analysis Gadget}
\usage{
farrell()
}
\value{
A data frame of the efficiency results including efficiency scores, peers' determination scale inefficiencies' evaluation and slacks' calculation.
}
\description{
farrell Data Envelopment Analysis Gadget
}
\examples{
if(interactive()){
farrell()
}


}
